+++ C#/unit_test/unit_test.performance.CGDK/CGDK.buffer.unit_test.performance.CGDK.projitems
@@ -10,5 +10,6 @@
   </PropertyGroup>
   <ItemGroup>
     <Compile Include="$(MSBuildThisFileDirectory)unit_test.performance.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)unit_test.performance.ex.cs" />
   </ItemGroup>
 </Project>
